backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
pillow==11.1.0
torch==2.4.1+cu121
transformers==4.51.3
diffusers==0.36.0
accelerate==1.5.2
celery==5.4.0
redis==5.2.1
python-dotenv==1.0.1
aiofiles==24.1.0
zipfile36==0.1.3
bitsandbytes==0.45.2
